druntime: Update to `wasi-sdk` 34

+++ druntime/src/core/stdc/stdint.d
@@ -400,8 +400,8 @@ else version (WASI)
 
     alias int_fast8_t   = byte;   ///
     alias uint_fast8_t  = ubyte;  ///
-    alias int_fast16_t  = short;  ///
-    alias uint_fast16_t = ushort; ///
+    alias int_fast16_t  = int;    ///
+    alias uint_fast16_t = uint;   ///
     alias int_fast32_t  = int;    ///
     alias uint_fast32_t = uint;   ///
